+++ packages/npm/oa42-generator/src/generators/files/client-ts.ts
@@ -3,7 +3,7 @@ import { RouterMode } from "goodrouter";
 import * as models from "../../models/index.js";
 import { packageInfo, toCamel, toPascal } from "../../utils/index.js";
 import { itt } from "../../utils/iterable-text-template.js";
-import { generateClientOperationFunctionBody } from "../bodies/index.js";
+import { generateClientOperationFunctionBody } from "../functions/index.js";
 import {
   generateOperationCredentialsType,
   generateOperationIncomingResponseType,
