+++ packages/npm/oa42-generator/src/generators/files/parameters-ts.ts
@@ -5,7 +5,7 @@ import { itt } from "../../utils/iterable-text-template.js";
 import {
   generateIsRequestParametersFunctionBody,
   generateIsResponseParametersFunctionBody,
-} from "../bodies/index.js";
+} from "../functions/index.js";
 import {
   generateOperationParametersTypes,
   generateOperationResultParameterTypes,
