+++ packages/npm/oa42-generator/src/generators/functions/client-operation.ts
