+++ packages/npm/oa42-generator/src/generators/functions/common-route-handler-method.ts
