+++ packages/npm/oa42-generator/src/generators/functions/endpoint-handler-method.ts
