+++ packages/npm/oa42-generator/src/generators/functions/is-response-parameters.ts
