+++ packages/npm/oa42-generator/src/generators/functions/route-handler-method.ts
